feat: prepare quotes for backtest

+++ internal/domain/quote/model.go
@@ -10,5 +10,5 @@ type Quote struct {
 	High   float64
 	Low    float64
 	Close  float64
-	Volume float32
+	Volume int32
 }
